requirements.txt
joblib
librosa
matplotlib
numpy
Pillow
PyYAML
soundfile
torch
transformers
